+++ apps/charterafrica/src/pages/api/v1/cron/[frequency].page.js
@@ -1,0 +1,34 @@
+import syncEcosystem from "@/charterafrica/lib/tools/syncEcosystem";
+
+const isApiKeyValid = (key) => {
+  return key && key === process.env.LOCAL_API_KEY;
+};
+
+const hourly = async () => {
+  return syncEcosystem(false);
+};
+
+const daily = () => {
+  return syncEcosystem(true);
+};
+
+const frequencyMap = {
+  hourly,
+  daily,
+};
+
+export default async function handler(req, res) {
+  const {
+    query: { frequency },
+  } = req;
+  const key = req.headers["x-api-key"];
+  if (!isApiKeyValid(key)) {
+    return res.status(403).json({ message: "INVALID API KEY" });
+  }
+  const func = frequencyMap[frequency];
+  if (func) {
+    const response = await func();
+    return res.status(200).json(response);
+  }
+  return res.status(404).json({ message: "UNKNOWN_FREQUENCY", frequency });
+}
